Nicolau

+++ src/Objeto.java
@@ -1,17 +1,16 @@
-
 public class Objeto {
 	
 	Ponto[] visao; // Pontos do objeto em coordenadas de vista
 	Ponto[] screen; // Pontos do objeto em coordenadas de tela (i.e., 2D)
 
-	double[][] triangulos;
+	int[][] triangulos;
 	
 	/*Pré-processamento*/
 	Ponto[] normais_triangulos;
 	Ponto[] normais_vertices;
 	//
 
-	public Objeto(Ponto[] visao, Ponto[] screen, Ponto[] normais_triangulos, Ponto[] normais_vertices, double[][] triangulos) {
+	public Objeto(Ponto[] visao, Ponto[] screen, Ponto[] normais_triangulos, Ponto[] normais_vertices, int[][] triangulos) {
 		this.visao = visao;
 		this.screen = screen;
 		this.normais_triangulos = normais_triangulos;
